+++ Examples/MovieProject/MovieProject.Logic/DTO/Media.cs
@@ -7,5 +7,14 @@
         public string Year { get; set; }
         public string Plot { get; set; }
         public string Runtime { get; set; }
+        public Languages Language { get; set; }
+
+        public enum Languages
+        {
+            English,
+            Spanish,
+            French,
+            Other
+        }
     }
 }