+++ Tool/SystemTestingTools/MockInstrumentation.cs
@@ -1,5 +1,7 @@
 ﻿using Microsoft.AspNetCore.Http;
 using System.Collections.Generic;
+using Microsoft.AspNetCore.Http.Extensions;
+using System.Reflection;
 
 namespace SystemTestingTools
 {
@@ -20,6 +22,29 @@ namespace SystemTestingTools
             if (context?.HttpContext?.Request?.Headers == null) return null;
 
             return context.HttpContext.Request.Headers[Constants.headerName];
+        }
+
+        internal static string GetUrl()
+        {
+            // not in a seession, could be happening during app start up or scheduled process
+            if (context?.HttpContext?.Request == null) return "No httpcontext available";
+
+            return $"{context.HttpContext.Request.Method} {context.HttpContext.Request.GetDisplayUrl()}";
+        }
+
+        private static readonly object padlock = new object();
+        private static string AppNameAndVersion = null;
+        internal static string GetAppNameAndVersion()
+        {
+            if (AppNameAndVersion == null)
+                lock (padlock)
+                    if (AppNameAndVersion == null) // double lock for the win :)
+                    {
+                        var assembly = Assembly.GetEntryAssembly().GetName();
+                        AppNameAndVersion = $"{assembly.Name} {assembly.Version}";
+                    }
+
+            return AppNameAndVersion;
         }
 
         /// <summary>